requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.5
numpy>=1.23.0
python-dotenv>=0.21.0
schedule>=1.2.0
requests>=2.28.0
openpyxl>=3.0.0
SQLAlchemy>=1.4.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
locust>=2.0.0